+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for `Pipfile.lock` (JSON) format in `DependencyParser` and `MultiScanner`
+
 ## [0.1.0] - 2026-09-09
 
 ### Added